src/test/java/kg/alatoo/notesapplication/controllers/NoteControllerTest.java
package kg.alatoo.notesapplication.controllers;

import kg.alatoo.notesapplication.dto.NoteDTO;
import kg.alatoo.notesapplication.services.NoteService;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.boot.test.mock.mockito.MockBean;
import org.springframework.http.MediaType;
import org.springframework.test.context.junit4.SpringRunner;
import org.springframework.test.web.servlet.MockMvc;

import static org.mockito.BDDMockito.given;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;

@RunWith(SpringRunner.class)
@WebMvcTest(NoteController.class)
public class NoteControllerTest {

    @Autowired
    private MockMvc mockMvc;

    @MockBean
    private NoteService noteService;

    @Test
    public void whenFindNoteById_thenReturnsNoteDTO() throws Exception {
        NoteDTO noteDTO = new NoteDTO();
        noteDTO.setId(1L);
        noteDTO.setTitle("Test Note");
        noteDTO.setContent("Test content");

        given(noteService.findById(1L)).willReturn(noteDTO);

        mockMvc.perform(get("/api/notes/1"))
                .andExpect(status().isOk())
                .andExpect(content().contentType(MediaType.APPLICATION_JSON))
                .andExpect(jsonPath("$.id").value(1))
                .andExpect(jsonPath("$.title").value("Test Note"))
                .andExpect(jsonPath("$.content").value("Test content"));
    }
}
